WoT restructure B2: split rangeQuery/renderedRange.ts into port interfaces + virtualRange.ts adapter
